day11: Complete

+++ src/day11/ThrowAction.kt
@@ -4,7 +4,7 @@ package day11
  * Algorithm for monkey to determine which other monkey to throw item to
  * according to divisibility test.
  */
-data class ThrowAction(val testDivisibleBy: Int, val trueMonkeyTarget: Int, val falseMonkeyTarget: Int) {
+data class ThrowAction(val testDivisibleBy: Long, val trueMonkeyTarget: Int, val falseMonkeyTarget: Int) {
     companion object {
         val testDivisibleByPattern = """Test: divisible by (\d+)""".toRegex()
         val trueMonkeyTargetPattern = """If true: throw to monkey (\d+)""".toRegex()
@@ -24,14 +24,14 @@ data class ThrowAction(val testDivisibleBy: Int, val trueMonkeyTarget: Int, val 
                 .matchEntire(falseMonkeyText)
                 ?.destructured
                 ?: throw IllegalArgumentException("invalid pattern: $falseMonkeyText")
-            return ThrowAction(testDivisibleBy.toInt(), trueMonkeyTarget.toInt(), falseMonkeyTarget.toInt())
+            return ThrowAction(testDivisibleBy.toLong(), trueMonkeyTarget.toInt(), falseMonkeyTarget.toInt())
         }
     }
 
     /**
      * Finds the next monkey number based on the given worry level [item].
      */
-    fun findNextMonkey(item: Int): Int = if (item % this.testDivisibleBy == 0) {
+    fun findNextMonkey(item: Long): Int = if (item % this.testDivisibleBy == 0L) {
         this.trueMonkeyTarget
     } else {
         this.falseMonkeyTarget
